rename: Community.tsx -> index.tsx
